+++ packages/commerce-sdk-react/src/hooks/useMutation.ts
@@ -19,11 +19,13 @@ import {
     ApiOptions,
     CacheUpdateGetter,
     MergedOptions,
-    OptionalCustomEndpointClientConfig
+    OptionalCustomEndpointClientConfig,
+    TMutationVariables
 } from './types'
 import {useAuthorizationHeader} from './useAuthorizationHeader'
 import useCustomerId from './useCustomerId'
 import {mergeOptions, updateCache} from './utils'
+import {handleInvalidToken, generateCustomEndpointOptions} from './helpers'
 
 /**
  * Helper for mutation hooks, contains most of the logic in order to keep individual hooks small.
@@ -54,12 +56,6 @@ export const useMutation = <
     })
 }
 
-type TMutationVariables = {
-    body?: unknown
-    parameters?: {[key: string]: string | number | boolean | string[] | number[]}
-    headers?: {[key: string]: string}
-} | void
-
 /**
  * A hook for SCAPI custom endpoint mutations.
  *
@@ -76,40 +72,21 @@ export const useCustomMutation = <TData = unknown, TError = unknown>(
 ) => {
     const auth = useAuthContext()
     const globalConfig = useConfig()
-    const globalHeaders = globalConfig.headers || {}
-    const globalClientConfig = {
-        parameters: {
-            clientId: globalConfig.clientId,
-            siteId: globalConfig.siteId,
-            organizationId: globalConfig.organizationId,
-            shortCode: globalConfig.shortCode
-        },
-        proxy: globalConfig.proxy
-    }
+    const logger = globalConfig.logger || console
 
     const createMutationFnWithAuth = (): MutationFunction<TData, TMutationVariables> => {
         return async (args): Promise<TData> => {
             const {access_token} = await auth.ready()
-            return (await helpers.callCustomEndpoint({
-                ...apiOptions,
-                options: {
-                    ...apiOptions.options,
-                    headers: {
-                        Authorization: `Bearer ${access_token}`,
-                        // Note the order of the following destructred objects is important.
-                        // Priority assending order: global config < mutation config < mutate func args
-                        ...globalHeaders,
-                        ...apiOptions.options?.headers,
-                        ...(args?.headers ? args.headers : {})
-                    },
-                    ...(args?.body ? {body: args.body} : {}),
-                    ...(args?.parameters ? {parameters: args.parameters} : {})
-                },
-                clientConfig: {
-                    ...globalClientConfig,
-                    ...(apiOptions.clientConfig || {})
-                }
-            })) as TData
+            return (await helpers
+                .callCustomEndpoint(
+                    generateCustomEndpointOptions(apiOptions, globalConfig, access_token, args)
+                )
+                .catch(async (error) => {
+                    const {access_token} = await handleInvalidToken(error, auth, logger)
+                    return await helpers.callCustomEndpoint(
+                        generateCustomEndpointOptions(apiOptions, globalConfig, access_token, args)
+                    )
+                })) as TData
         }
     }